+++ akshare/__init__.py
@@ -3204,9 +3204,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.17.95 fix: fix stock_info_sh_name_code interface
 1.17.96 fix: fix stock_info_sh_name_code interface
 1.17.97 fix: fix stock_financial_debt_new_ths interface
+1.17.98 fix: fix stock_jgdy_tj_em interface
 """
 
-__version__ = "1.17.97"
+__version__ = "1.17.98"
 __author__ = "AKFamily"
 
 import sys